Split index.ts into route modules and services
Extract 1,000+ lines from index.ts (1,245 → 240 lines) into:
- src/routes/ (today, program, setup, progress, routine, sync, webhooks)
- src/services/ (activate-program, sync)
- src/utils/sse.ts (sseErrorCard helper)

Move Env interface to types.ts, loadProgram to queries.ts.
Pure refactoring — no behavior changes.

Closes #23
